requirements.txt
arabic_reshaper==3.0.0
mysql_connector_repackaged==0.3.1
nest_asyncio==1.6.0
pandas==2.2.2
Pillow==10.4.0
python-telegram-bot==21.5
